+++ packages/scalar.aspnetcore/src/DefaultHttpClient.cs
@@ -1,0 +1,8 @@
+namespace Scalar.AspNetCore;
+
+internal sealed class DefaultHttpClient
+{
+    public required string TargetKey { get; init; }
+    
+    public required string ClientKey { get; init; }
+}
